+++ core/src/formatter.rs
@@ -1007,7 +1007,7 @@ mod tests {
             .parser(DelphiLogicalLineParser {})
             .file_formatter(IndentBasedOnLineNumber {})
             .reconstructor(DelphiLogicalLinesReconstructor::new(
-                ReconstructionSettings::new("\n", " ", " ").unwrap(),
+                ReconstructionSettings::new(LineEnding::Lf, TabKind::Soft, 1, 1),
             ))
             .build();
         run_test(
@@ -1042,7 +1042,7 @@ mod tests {
             .file_formatter(IndentBasedOnLineNumber {})
             .file_formatter(IndentSecondLine3SpacesIfNoNewLine {})
             .reconstructor(DelphiLogicalLinesReconstructor::new(
-                ReconstructionSettings::new("\n", " ", " ").unwrap(),
+                ReconstructionSettings::new(LineEnding::Lf, TabKind::Soft, 1, 1),
             ))
             .build();
         run_test(
@@ -1095,7 +1095,7 @@ mod tests {
             .file_formatter(IndentSecondLine3SpacesIfNoNewLine {})
             .line_formatter(RetainSpacesLogcialLinesOnNewLines {})
             .reconstructor(DelphiLogicalLinesReconstructor::new(
-                ReconstructionSettings::new("\n", " ", " ").unwrap(),
+                ReconstructionSettings::new(LineEnding::Lf, TabKind::Soft, 1, 1),
             ))
             .build();
         run_test(