+++ core/src/lang.rs
@@ -1,4 +1,4 @@
-use std::{borrow::Cow, error::Error, fmt::Display, vec::Drain};
+use std::{borrow::Cow, vec::Drain};
 
 use crate::prelude::*;
 
@@ -579,61 +579,53 @@ impl<'a> From<LogicalLines<'a>> for (&'a [Token<'a>], Vec<LogicalLine>) {
     }
 }
 
-#[derive(Debug, PartialEq, Eq)]
-pub struct InvalidReconstructionSettingsError {
-    msg: String,
+#[derive(Copy, Clone, Debug)]
+pub enum LineEnding {
+    Crlf,
+    Lf,
 }
 
-impl Display for InvalidReconstructionSettingsError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "invalid reconstruction settings: {}", self.msg)
-    }
-}
-
-impl Error for InvalidReconstructionSettingsError {}
-
-impl InvalidReconstructionSettingsError {
-    fn new<S: Into<String>>(msg: S) -> Self {
-        InvalidReconstructionSettingsError { msg: msg.into() }
-    }
+#[derive(Copy, Clone)]
+pub enum TabKind {
+    Soft,
+    Hard,
 }
 
 #[derive(Clone)]
 pub struct ReconstructionSettings {
-    newline_str: String,
+    newline_str: &'static str,
     indentation_str: String,
     continuation_str: String,
 }
 impl ReconstructionSettings {
-    pub fn new<S: Into<String> + AsRef<str>>(
-        newline_str: S,
-        indentation_str: S,
-        continuation_str: S,
-    ) -> Result<Self, InvalidReconstructionSettingsError> {
-        for (name, val) in &[
-            ("newline", newline_str.as_ref()),
-            ("indentation", indentation_str.as_ref()),
-            ("continuation", continuation_str.as_ref()),
-        ] {
-            if val.is_empty() {
-                return Err(InvalidReconstructionSettingsError::new(format!(
-                    "{name} sequence cannot be blank"
-                )));
-            } else if count_leading_whitespace(val) != val.len() {
-                return Err(InvalidReconstructionSettingsError::new(format!(
-                    "{name} sequence must be all whitespace (was {val:?})"
-                )));
-            }
-        }
+    pub fn new(
+        line_ending: LineEnding,
+        tab: TabKind,
+        indent_width: u8,
+        continuation_width: u8,
+    ) -> Self {
+        let newline_str = match line_ending {
+            LineEnding::Crlf => "\r\n",
+            LineEnding::Lf => "\n",
+        };
 
-        Ok(ReconstructionSettings {
-            newline_str: newline_str.into(),
-            indentation_str: indentation_str.into(),
-            continuation_str: continuation_str.into(),
-        })
+        let indent = match tab {
+            TabKind::Soft => " ",
+            TabKind::Hard => "\t",
+        };
+
+        let indentation_str = indent.repeat(indent_width.into());
+        let continuation_str = indent.repeat(continuation_width.into());
+
+        Self {
+            newline_str,
+            indentation_str,
+            continuation_str,
+        }
     }
+
     pub fn get_newline_str(&self) -> &str {
-        &self.newline_str
+        self.newline_str
     }
     pub fn get_indentation_str(&self) -> &str {
         &self.indentation_str
@@ -764,87 +756,6 @@ impl LogicalLineFileFormatter for FormatterKind {
                 .iter()
                 .for_each(|logical_line| formatter.format(formatted_tokens, logical_line)),
             FormatterKind::FileFormatter(formatter) => formatter.format(formatted_tokens, input),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    mod reconstruction_validation {
-        use super::*;
-
-        #[test]
-        fn normal_whitespace() {
-            assert!(ReconstructionSettings::new("\n", "\t", "    ",).is_ok());
-        }
-
-        #[test]
-        fn weird_whitespace() {
-            assert!(ReconstructionSettings::new("\0", "\x1F", "\u{3000}",).is_ok());
-        }
-
-        #[test]
-        fn borrowed_or_owned() {
-            assert!(ReconstructionSettings::new("\n", " ", " ").is_ok());
-            assert!(
-                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned())
-                    .is_ok()
-            );
-        }
-
-        #[test]
-        fn invalid_empty() {
-            use super::InvalidReconstructionSettingsError as E;
-
-            assert_eq!(
-                ReconstructionSettings::new("", "", "").err(),
-                Some(E::new("newline sequence cannot be blank"))
-            );
-            assert_eq!(
-                ReconstructionSettings::new("\n", "", "").err(),
-                Some(E::new("indentation sequence cannot be blank"))
-            );
-            assert_eq!(
-                ReconstructionSettings::new("\n", " ", "").err(),
-                Some(E::new("continuation sequence cannot be blank"))
-            );
-        }
-
-        #[test]
-        fn invalid_non_whitespace() {
-            use super::InvalidReconstructionSettingsError as E;
-
-            assert_eq!(
-                ReconstructionSettings::new("a", " ", " ").err(),
-                Some(E::new(
-                    "newline sequence must be all whitespace (was \"a\")"
-                ))
-            );
-            assert_eq!(
-                ReconstructionSettings::new("\n", "a", " ").err(),
-                Some(E::new(
-                    "indentation sequence must be all whitespace (was \"a\")"
-                ))
-            );
-            assert_eq!(
-                ReconstructionSettings::new("\n", " ", "a").err(),
-                Some(E::new(
-                    "continuation sequence must be all whitespace (was \"a\")"
-                ))
-            );
-        }
-
-        #[test]
-        fn display() {
-            assert_eq!(
-                ReconstructionSettings::new("", " ", " ")
-                    .err()
-                    .unwrap()
-                    .to_string(),
-                "invalid reconstruction settings: newline sequence cannot be blank"
-            );
         }
     }
 }